+++ packages/medusa/src/api/store/orders/query-config.ts
@@ -69,6 +69,7 @@ export const defaultStoreRetrieveOrderFields = [
 export const retrieveTransformQueryConfig = {
   defaults: defaultStoreRetrieveOrderFields,
   disallowed: disallowedStorePivotFields,
+  storeRelationsLimit: 3,
   isList: false,
 }